requirements.txt
streamlit
Pillow
numpy
opencv-python-headless
pandas
scikit-learn
joblib
